Swapping Long to BigDecimal
* Swapped `Long` for `BigDecimal` in places that will need it, specifically for large nhash quantities
* Should fix the Rich List issue
* Now rounding circulating supply, community pool values as those are whole `nhash` values

closes: #380
closes: #381

+++ service/src/main/kotlin/io/provenance/explorer/domain/entities/Validators.kt
@@ -35,6 +35,7 @@ import org.jetbrains.exposed.sql.leftJoin
 import org.jetbrains.exposed.sql.select
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.joda.time.DateTime
+import java.math.BigDecimal
 import java.sql.ResultSet
 
 object ValidatorsCacheTable : CacheIdTable<Int>(name = "validators_cache") {
@@ -330,7 +331,7 @@ object ValidatorMarketRateTable : IntIdTable(name = "validator_market_rate") {
 
 class ValidatorMarketRateRecord(id: EntityID<Int>) : IntEntity(id) {
     companion object : IntEntityClass<ValidatorMarketRateRecord>(ValidatorMarketRateTable) {
-        fun buildInsert(txInfo: TxData, proposer: String, tx: ServiceOuterClass.GetTxResponse, msgBasedFees: Long) =
+        fun buildInsert(txInfo: TxData, proposer: String, tx: ServiceOuterClass.GetTxResponse, msgBasedFees: BigDecimal) =
             listOf(
                 0,
                 txInfo.blockHeight,
